+++ server/db/update_scripts/22_update_v6.7.0_v6.8.0.sql
@@ -68,8 +68,10 @@ INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `
 INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`)  VALUES (get_style_id('formUserInputLog'), get_field_id('label_cancel'), '', 'Cancel button label on the confirmation modal');
 INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`)  VALUES (get_style_id('formUserInputLog'), get_field_id('label_continue'), 'OK', 'Continue button for the modal when the button is clicked');
 INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`)  VALUES (get_style_id('formUserInputLog'), get_field_id('label_message'), 'Do you want to continue?', 'The message shown on the modal');
+INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`)  VALUES (get_style_id('formUserInputLog'), get_field_id('url_cancel'), '', 'The target URL of the cancel button.');
 
 INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`)  VALUES (get_style_id('formUserInputRecord'), get_field_id('confirmation_title'), '', 'Confirmation title for the modal when the button is clicked');
 INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`)  VALUES (get_style_id('formUserInputRecord'), get_field_id('label_cancel'), '', 'Cancel button label on the confirmation modal');
 INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`)  VALUES (get_style_id('formUserInputRecord'), get_field_id('label_continue'), 'OK', 'Continue button for the modal when the button is clicked');
 INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`)  VALUES (get_style_id('formUserInputRecord'), get_field_id('label_message'), 'Do you want to continue?', 'The message shown on the modal');
+INSERT IGNORE INTO `styles_fields` (`id_styles`, `id_fields`, `default_value`, `help`)  VALUES (get_style_id('formUserInputRecord'), get_field_id('url_cancel'), '', 'The target URL of the cancel button.');